third_party/blink/web_tests/virtual/mutation-events-enabled/dom/mutation-event-tests/fast/forms/select/add-and-remove-option-expected.txt

CONSOLE WARNING: Listener added for a 'DOMNodeInserted' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeInserted' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeInserted' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
Test for bug 13287: Cannot change SELECT to a dynamically created option.

Test that DOM is updated before DOMNodeInserted is dispatched.

1
2
 
1
2
 
1
2

Passed
Passed
Passed
